requirements.txt
Django==3.2.10
amqp==5.0.9
bcoding==1.5
celery==5.2.3
django-cors-headers==3.10.1
django-jsonify==0.3.0
#importlib==1.0.4 ## no longer needed
isodate==0.6.1
oauth2==1.9.0.post1
psycopg2==2.9.3
pycryptodome==3.12.0
#pyinotify ## not used apparently and is linux-only
python-jose==3.3.0
pytz==2021.3
requests==2.26.0
rfc3987==1.3.8
supervisor==4.2.3